Security cleanup.
