+++ dcc-portal-server/src/main/java/org/icgc/dcc/portal/server/health/ICGCHealthIndicator.java
@@ -19,7 +19,7 @@ package org.icgc.dcc.portal.server.health;
 import static com.google.common.base.Strings.isNullOrEmpty;
 import static org.icgc.dcc.common.client.api.daco.DACOClient.UserType.CUD;
 
-import org.icgc.dcc.portal.server.service.AuthService;
+import org.icgc.dcc.portal.server.security.AuthService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.health.AbstractHealthIndicator;
 import org.springframework.boot.actuate.health.Health.Builder;
