+++ dcc-portal-server/src/main/java/org/icgc/dcc/portal/server/resource/core/DownloadResource.java
@@ -52,7 +52,6 @@ import org.icgc.dcc.download.core.jwt.JwtService;
 import org.icgc.dcc.download.core.model.DownloadFile;
 import org.icgc.dcc.download.core.model.JobUiInfo;
 import org.icgc.dcc.download.core.response.JobResponse;
-import org.icgc.dcc.portal.server.auth.Auth;
 import org.icgc.dcc.portal.server.config.ServerProperties.DownloadProperties;
 import org.icgc.dcc.portal.server.download.DownloadDataTypes;
 import org.icgc.dcc.portal.server.download.DownloadResources;
@@ -63,6 +62,7 @@ import org.icgc.dcc.portal.server.model.Query;
 import org.icgc.dcc.portal.server.model.User;
 import org.icgc.dcc.portal.server.model.param.FiltersParam;
 import org.icgc.dcc.portal.server.resource.Resource;
+import org.icgc.dcc.portal.server.security.jersey.Auth;
 import org.icgc.dcc.portal.server.service.DonorService;
 import org.icgc.dcc.portal.server.service.ForbiddenAccessException;
 import org.icgc.dcc.portal.server.service.NotFoundException;
