+++ dcc-portal-server/src/main/java/org/icgc/dcc/portal/server/resource/core/TokenResource.java
@@ -34,12 +34,12 @@ import javax.ws.rs.Produces;
 import javax.ws.rs.core.Response;
 
 import org.hibernate.validator.constraints.NotEmpty;
-import org.icgc.dcc.portal.server.auth.Auth;
 import org.icgc.dcc.portal.server.model.AccessToken;
 import org.icgc.dcc.portal.server.model.AccessTokenScopes;
 import org.icgc.dcc.portal.server.model.Tokens;
 import org.icgc.dcc.portal.server.model.User;
 import org.icgc.dcc.portal.server.resource.Resource;
+import org.icgc.dcc.portal.server.security.jersey.Auth;
 import org.icgc.dcc.portal.server.service.TokenService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
