+++ dcc-portal-server/src/main/java/org/icgc/dcc/portal/server/resource/security/AuthResource.java
@@ -24,10 +24,10 @@ import static javax.ws.rs.core.HttpHeaders.SET_COOKIE;
 import static javax.ws.rs.core.Response.status;
 import static javax.ws.rs.core.Response.Status.NOT_MODIFIED;
 import static javax.ws.rs.core.Response.Status.OK;
-import static org.icgc.dcc.portal.server.util.AuthUtils.createSessionCookie;
-import static org.icgc.dcc.portal.server.util.AuthUtils.deleteCookie;
-import static org.icgc.dcc.portal.server.util.AuthUtils.stringToUuid;
-import static org.icgc.dcc.portal.server.util.AuthUtils.throwAuthenticationException;
+import static org.icgc.dcc.portal.server.security.AuthUtils.createSessionCookie;
+import static org.icgc.dcc.portal.server.security.AuthUtils.deleteCookie;
+import static org.icgc.dcc.portal.server.security.AuthUtils.stringToUuid;
+import static org.icgc.dcc.portal.server.security.AuthUtils.throwAuthenticationException;
 
 import java.util.AbstractMap.SimpleImmutableEntry;
 import java.util.Collection;
@@ -52,7 +52,7 @@ import org.icgc.dcc.portal.server.config.ServerProperties.AuthProperties;
 import org.icgc.dcc.portal.server.config.ServerProperties.CrowdProperties;
 import org.icgc.dcc.portal.server.model.User;
 import org.icgc.dcc.portal.server.resource.Resource;
-import org.icgc.dcc.portal.server.service.AuthService;
+import org.icgc.dcc.portal.server.security.AuthService;
 import org.icgc.dcc.portal.server.service.CmsAuthService;
 import org.icgc.dcc.portal.server.service.SessionService;
 import org.springframework.beans.factory.annotation.Autowired;
