+++ dcc-portal-server/src/main/java/org/icgc/dcc/portal/server/security/AuthenticationException.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2016 The Ontario Institute for Cancer Research. All rights reserved.                             
+ * Copyright (c) 2014 The Ontario Institute for Cancer Research. All rights reserved.                             
  *                                                                                                               
  * This program and the accompanying materials are made available under the terms of the GNU Public License v3.0.
  * You should have received a copy of the GNU General Public License along with                                  
@@ -15,32 +15,23 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN                         
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.icgc.dcc.portal.server.auth;
+package org.icgc.dcc.portal.server.security;
 
-import com.google.common.base.Optional;
+import java.net.URI;
+import java.util.Collection;
 
-/**
- * An interface for classes which authenticate user-provided credentials and return principal objects.
- *
- * @param <C> the type of credentials the authenticator can authenticate
- * @param <P> the type of principals the authenticator returns
- */
-public interface Authenticator<C, P> {
+import lombok.EqualsAndHashCode;
+import lombok.NonNull;
+import lombok.Value;
 
-  /**
-   * Given a set of user-provided credentials, return an optional principal.
-   *
-   * <p>
-   * If the credentials are valid and map to a principal, returns an {@code Optional.of(p)}.
-   * </p>
-   *
-   * <p>
-   * If the credentials are invalid, returns an {@code Optional.absent()}.
-   * </p>
-   *
-   * @param credentials a set of user-provided credentials
-   * @return either an authenticated principal or an absent optional
-   * @throws AuthenticationException if the credentials cannot be authenticated due to an underlying error
-   */
-  Optional<P> authenticate(C credentials) throws AuthenticationException;
+@Value
+@EqualsAndHashCode(callSuper = false)
+public class AuthenticationException extends RuntimeException {
+
+  @NonNull
+  String message;
+  URI redirect;
+  @NonNull
+  Collection<String> invalidCookies;
+
 }
