+++ dcc-portal-server/src/main/java/org/icgc/dcc/portal/server/security/jersey/Auth.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2015 The Ontario Institute for Cancer Research. All rights reserved.                             
+ * Copyright (c) 2016 The Ontario Institute for Cancer Research. All rights reserved.                             
  *                                                                                                               
  * This program and the accompanying materials are made available under the terms of the GNU Public License v3.0.
  * You should have received a copy of the GNU General Public License along with                                  
@@ -15,34 +15,28 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN                         
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.icgc.dcc.portal.server.auth.oauth;
+package org.icgc.dcc.portal.server.security.jersey;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import com.fasterxml.jackson.annotation.JsonProperty;
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
 
-import lombok.Value;
+/**
+ * This annotation is used to inject authenticated principal objects into protected JAX-RS resource methods.
+ *
+ * @see Authenticator
+ */
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ ElementType.PARAMETER, ElementType.FIELD })
+public @interface Auth {
 
-@Value
-class AccessTokenResponse {
-
-  String id;
-  String type;
-  String description;
-  int expiresIn;
-  String scope;
-
-  @JsonCreator
-  public AccessTokenResponse(
-      @JsonProperty("access_token") String id,
-      @JsonProperty("token_type") String type,
-      @JsonProperty("desc") String description,
-      @JsonProperty("expires_in") int expires,
-      @JsonProperty("scope") String scope) {
-    this.id = id;
-    this.type = type;
-    this.description = description;
-    this.expiresIn = expires;
-    this.scope = scope;
-  }
+  /**
+   * If {@code true}, the request will not be processed in the absence of a valid principal. If {@code false},
+   * {@code null} will be passed in as a principal. Defaults to {@code true}.
+   */
+  boolean required() default true;
 
 }
