+++ dcc-portal-server/src/main/java/org/icgc/dcc/portal/server/security/oauth/CheckTokenResponse.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2013 The Ontario Institute for Cancer Research. All rights reserved.                             
+ * Copyright (c) 2015 The Ontario Institute for Cancer Research. All rights reserved.                             
  *                                                                                                               
  * This program and the accompanying materials are made available under the terms of the GNU Public License v3.0.
  * You should have received a copy of the GNU General Public License along with                                  
@@ -15,41 +15,32 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN                         
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.icgc.dcc.portal.server.auth;
+package org.icgc.dcc.portal.server.security.oauth;
 
-import java.util.Optional;
-import java.util.UUID;
+import java.util.List;
 
-import lombok.NonNull;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonProperty;
+
 import lombok.Value;
 
-/**
- * Value object to provide the following to {@link UserAuthenticator}:
- * <ul>
- * <li>Storage of the necessary credentials for OpenID authentication</li>
- * </ul>
- */
 @Value
-public class UserCredentials {
+@JsonIgnoreProperties(ignoreUnknown = true)
+class CheckTokenResponse {
 
-  /**
-   * The session token acting as a surrogate for the OpenID token.
-   */
-  @NonNull
-  private final Optional<UUID> sessionToken;
+  int exp;
+  String desc;
+  List<String> scope;
 
-  /**
-   * OAuth {@code portal.download} token.
-   */
-  @NonNull
-  private final Optional<String> accessToken;
-
-  public boolean isWebSession() {
-    return sessionToken.isPresent();
-  }
-
-  public boolean isAPI() {
-    return accessToken.isPresent();
+  @JsonCreator
+  public CheckTokenResponse(
+      @JsonProperty("exp") int exp,
+      @JsonProperty("desc") String desc,
+      @JsonProperty("scope") List<String> scope) {
+    this.exp = exp;
+    this.desc = desc;
+    this.scope = scope;
   }
 
 }
