+++ dcc-portal-server/src/test/java/org/icgc/dcc/portal/server/resource/core/DownloadResourceTest.java
@@ -44,15 +44,15 @@ import org.icgc.dcc.download.core.model.DownloadFile;
 import org.icgc.dcc.download.core.model.DownloadFileType;
 import org.icgc.dcc.download.core.model.JobUiInfo;
 import org.icgc.dcc.download.core.response.JobResponse;
-import org.icgc.dcc.portal.server.auth.UserAuthProvider;
-import org.icgc.dcc.portal.server.auth.UserAuthenticator;
-import org.icgc.dcc.portal.server.auth.oauth.OAuthClient;
 import org.icgc.dcc.portal.server.config.ServerProperties.CrowdProperties;
 import org.icgc.dcc.portal.server.config.ServerProperties.DownloadProperties;
 import org.icgc.dcc.portal.server.download.JobInfo;
 import org.icgc.dcc.portal.server.jersey.mapper.BadRequestExceptionMapper;
 import org.icgc.dcc.portal.server.model.Query;
 import org.icgc.dcc.portal.server.model.User;
+import org.icgc.dcc.portal.server.security.jersey.UserAuthProvider;
+import org.icgc.dcc.portal.server.security.jersey.UserAuthenticator;
+import org.icgc.dcc.portal.server.security.oauth.OAuthClient;
 import org.icgc.dcc.portal.server.service.DonorService;
 import org.icgc.dcc.portal.server.service.ForbiddenAccessException;
 import org.icgc.dcc.portal.server.service.NotFoundException;
