+++ dcc-portal-server/src/test/java/org/icgc/dcc/portal/server/security/openid/DistributedNonceVerifierTest.java
@@ -15,10 +15,10 @@
  * IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN                         
  * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.icgc.dcc.portal.server.auth.openid;
+package org.icgc.dcc.portal.server.security.openid;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.icgc.dcc.portal.server.auth.openid.DistributedNonceVerifier.NONCE_CACHE_NAME;
+import static org.icgc.dcc.portal.server.security.openid.DistributedNonceVerifier.NONCE_CACHE_NAME;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 import static org.openid4java.consumer.NonceVerifier.INVALID_TIMESTAMP;
@@ -28,6 +28,7 @@ import static org.openid4java.consumer.NonceVerifier.TOO_OLD;
 
 import java.util.Date;
 
+import org.icgc.dcc.portal.server.security.openid.DistributedNonceVerifier;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
